requirements.txt
django==4.2.6
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2==2.9.9
django-solo==2.1.0
django-jazzmin==2.6.0
django-imagekit==5.0.0